RAG/RAG_Production_Bytewax_pipeline/pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["vaibhawkhemka <vaibhawkhemka6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.4"
pandas = "^2.2.2"
pydantic-settings = "^2.2.1"
qdrant-client = "^1.9.1"
sentence-transformers = "^2.7.0"
langchain-community = "^0.0.36"
langchain = "^0.1.17"
pinecone-client = "^4.0.0"
bytewax = "^0.19.1"
matplotlib = "^3.8.4"
umap = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
